+++ projects/ngx-webrtc-lib/src/lib/components/controls/controls.component.ts
@@ -1,10 +1,11 @@
-import { Component, ChangeDetectionStrategy, Input, Output, EventEmitter } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
 
 @Component({
   selector: 'ngx-controls',
   templateUrl: './controls.component.html',
   styleUrls: ['./controls.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
 })
 export class ControlsComponent {
   @Input() cameraEnabled = true;