+++ projects/ngx-webrtc-lib/src/lib/webrtc-lib.module.ts
@@ -1,32 +1,16 @@
 import { CommonModule } from '@angular/common';
 import { ModuleWithProviders, NgModule } from '@angular/core';
 
-import {
-  ControlsComponent,
-  SpinnerComponent,
-  TimerComponent,
-  VideoCallComponent,
-} from './components';
-import { WebRtcComponent } from './webrtc.component';
-
-import {
-  ToggleDirective,
-} from './directives';
-
 import { AgoraConfig } from './models';
-import { DialogService, VideoCallDialogService, WebRtcService } from './services';
+import { WebRtcService } from './services';
+import { WebRtcComponent } from './webrtc.component';
 
 @NgModule({
   declarations: [
-    ControlsComponent,
-    WebRtcComponent,
-    TimerComponent,
-    SpinnerComponent,
-    ToggleDirective,
-    VideoCallComponent,
   ],
   imports: [
     CommonModule,
+    WebRtcComponent,
   ],
   exports: [
     WebRtcComponent,
@@ -37,8 +21,6 @@ export class WebRtcModule {
     return {
       ngModule: WebRtcModule,
       providers: [
-        DialogService,
-        VideoCallDialogService,
         WebRtcService,
         {
           provide: 'AgoraConfig',