+++ projects/ngx-webrtc-lib/src/public-api.ts
@@ -4,5 +4,6 @@
 
 export * from './lib/models/call-dialog-data.model';
 export * from './lib/services/video-call-dialog.service';
-export * from './lib/webrtc.component';
 export * from './lib/webrtc-lib.module';
+export * from './lib/webrtc-lib.provider';
+export * from './lib/webrtc.component';