+++ src/app/app.component.ts
@@ -1,10 +1,13 @@
 import { ChangeDetectionStrategy, Component, HostBinding } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
 import packageJson from '../../package.json';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss'],
+  standalone: true,
+  imports: [RouterOutlet],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AppComponent {